Zapilena rabota postroitelä kart osvescenija. Pocineno nepraviljnoje vystavlenije nomerov tekstur dlä veršin krivyh. Dodelana zagruzka istocnikov osvescenija iz BSP. Popravleny šejdery.

+++ shaders/point_light_shadowmap_v.glsl
@@ -1,7 +1,7 @@
 in vec3 pos;
 in vec2 tex_coord;
 in vec2 lightmap_coord;
-in vec3 tex_maps;
+in ivec3 tex_maps;
 in vec3 normal;
 
 out vec3 g_pos;
