+++ shaders/secondary_light_pass_v.glsl
@@ -3,7 +3,7 @@ uniform vec4 clip_plane;
 in vec3 pos;
 in vec2 tex_coord;
 in vec2 lightmap_coord;
-in vec3 tex_maps;
+in ivec3 tex_maps;
 in vec3 normal;
 
 out vec3 g_tex_coord;
