+++ shaders/sun_shadowmap_v.glsl
@@ -4,7 +4,7 @@ uniform vec4 clip_planes[4];
 in vec3 pos;
 in vec2 tex_coord;
 in vec2 lightmap_coord;
-in vec3 tex_maps;
+in ivec3 tex_maps;
 in vec3 normal;
 
 void main()
